+++ server/src/entities/Message.ts
@@ -1,0 +1,29 @@
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  ManyToOne,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+} from "typeorm";
+import { Conversation } from "./Conversation";
+
+@Entity()
+export class Message {
+  @PrimaryGeneratedColumn("uuid")
+  uuid: string;
+
+  @Column()
+  content: string;
+
+  @CreateDateColumn()
+  createdAt: Date;
+
+  @UpdateDateColumn()
+  updatedAt: Date;
+
+  @ManyToOne(() => Conversation, (conversation) => conversation.messages, {
+    cascade: false,
+  })
+  conversation: Conversation;
+}